refactor expandable section into new component
