+++ frontend/src/components/expandable_area/stylesheet.styl
